Move manage comments JSP and action
