+++ study/src/org/labkey/study/StudyInternalServiceImpl.java
@@ -6,12 +6,15 @@ import org.labkey.api.security.User;
 import org.labkey.api.specimen.query.SpecimenQueryView;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.study.CohortFilter;
+import org.labkey.api.study.Dataset;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyInternalService;
+import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.model.ParticipantDataset;
 import org.labkey.api.study.model.ParticipantInfo;
 import org.labkey.api.view.ViewContext;
 import org.labkey.study.controllers.specimen.SpecimenUtils;
+import org.labkey.study.model.SecurityType;
 import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
 import org.springframework.validation.BindException;
@@ -79,8 +82,52 @@ public class StudyInternalServiceImpl implements StudyInternalService
     }
 
     @Override
+    public String getParticipantCommentProperty(Study study)
+    {
+        return null;
+    }
+
+    @Override
     public Integer getParticipantVisitCommentDatasetId(Study study)
     {
         return ((StudyImpl)study).getParticipantVisitCommentDatasetId();
+    }
+
+    @Override
+    public String getParticipantVisitCommentProperty(Study study)
+    {
+        return ((StudyImpl)study).getParticipantVisitCommentProperty();
+    }
+
+    @Override
+    public List<? extends Dataset> getDatasets(Study study)
+    {
+        return StudyManager.getInstance().getDatasetDefinitions(study);
+    }
+
+    @Override
+    public boolean hasEditableDatasets(Study study)
+    {
+        SecurityType securityType = ((StudyImpl)study).getSecurityType();
+        return securityType != SecurityType.ADVANCED_READ && securityType != SecurityType.BASIC_READ;
+    }
+
+    @Override
+    public void saveCommentsSettings(Study s, User user, Integer participantCommentDatasetId, String participantCommentProperty, Integer participantVisitCommentDatasetId, String participantVisitCommentProperty)
+    {
+        StudyImpl study = (StudyImpl)s;
+
+        // participant comment dataset
+        study.setParticipantCommentDatasetId(participantCommentDatasetId);
+        study.setParticipantCommentProperty(participantCommentProperty);
+
+        // participant/visit comment dataset
+        if (study.getTimepointType() != TimepointType.CONTINUOUS)
+        {
+            study.setParticipantVisitCommentDatasetId(participantVisitCommentDatasetId);
+            study.setParticipantVisitCommentProperty(participantVisitCommentProperty);
+        }
+
+        StudyManager.getInstance().updateStudy(user, study);
     }
 }
